config: support UEFI in serial images, off by default for now; closes #138
While here switch images to UFS2.

+++ build/dvd.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-# Copyright (c) 2014-2019 Franco Fichtner <franco@opnsense.org>
+# Copyright (c) 2014-2021 Franco Fichtner <franco@opnsense.org>
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -47,7 +47,8 @@ setup_mtree ${STAGEDIR}/work
 setup_entropy ${STAGEDIR}/work
 
 UEFIBOOT=
-if [ ${PRODUCT_ARCH} = "amd64" -a -n "${PRODUCT_UEFI}" ]; then
+if [ ${PRODUCT_ARCH} = "amd64" -a -n "${PRODUCT_UEFI}" -a \
+    -z "${PRODUCT_UEFI%%*dvd*}" ]; then
 	dd if=/dev/zero of=${STAGEDIR}/efiboot.img bs=4k count=200
 	DEV=$(mdconfig -a -t vnode -f ${STAGEDIR}/efiboot.img)
 	newfs_msdos -F 12 -m 0xf8 /dev/${DEV}
